requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.9.0
Django==5.2
django-anymail==13.0
django-appconf==1.1.0
django-environ==0.12.0
gunicorn==23.0.0
idna==3.10
joblib==1.4.2
kombu==5.5.3
nltk==3.9.1
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg==3.2.6
psycopg-pool==3.2.6
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
